+++ src/commands/activate.ts
@@ -5,23 +5,26 @@ import * as i18n from './../i18n';
 import * as outdatedMessage from './../messages/outdated';
 
 /** Activate the icon theme by changing the settings for the iconTheme. */
-export const activateIconTheme = () => {
+export const activateIcons = () => {
     if (!versioning.checkVersionSupport('1.10.0')) {
         outdatedMessage.showOutdatedMessage();
-        return Promise.reject('Outdated version of vscode!');
+        console.error('Outdated version of vscode!');
     }
     return setIconTheme();
 };
 
 /** Set the icon theme in the config. */
-const setIconTheme = () => {
+const setIconTheme = async () => {
     // global user config
-    return helpers.getConfig().update('workbench.iconTheme', 'material-icon-theme', true)
-        .then(() => {
-            // local workspace config
-            if (helpers.getConfig().inspect('workbench.iconTheme').workspaceValue !== undefined) {
-                helpers.getConfig().update('workbench.iconTheme', 'material-icon-theme');
-            }
-            vscode.window.showInformationMessage(i18n.translate('activated'));
-        });
+    try {
+        await helpers.getConfig().update('workbench.iconTheme', 'material-icon-theme', true);
+
+        // local workspace config
+        if (helpers.getConfig().inspect('workbench.iconTheme').workspaceValue !== undefined) {
+            helpers.getConfig().update('workbench.iconTheme', 'material-icon-theme');
+        }
+        vscode.window.showInformationMessage(i18n.translate('activated'));
+    } catch (error) {
+        console.error(error);
+    }
 };