+++ src/commands/index.ts
@@ -1,53 +1,27 @@
 import * as vscode from 'vscode';
-import { activateIconTheme } from './activate';
+import { activateIcons } from './activate';
 import { toggleFolderArrows } from './folderArrows';
 import { changeFolderColor } from './folderColor';
 import { changeFolderTheme } from './folders';
+import { toggleGrayscale } from './grayscale';
 import { toggleIconPacks } from './iconPacks';
 import { changeOpacity } from './opacity';
 import { restoreDefaultConfig } from './restoreConfig';
+import { changeSaturation } from './saturation';
 
-// Activate theme
-const activateThemeCommand = vscode.commands.registerCommand('material-icon-theme.activateIcons', () => {
-    activateIconTheme();
+const commands = {
+    activateIcons,
+    toggleIconPacks,
+    changeFolderTheme,
+    changeFolderColor,
+    restoreDefaultConfig,
+    toggleFolderArrows,
+    changeOpacity,
+    toggleGrayscale,
+    changeSaturation
+};
+
+export const registered = Object.keys(commands).map(commandName => {
+    const callCommand = () => commands[commandName]();
+    return vscode.commands.registerCommand(`material-icon-theme.${commandName}`, callCommand);
 });
-
-// Icon packs
-const toggleIconPacksCommand = vscode.commands.registerCommand('material-icon-theme.toggleIconPacks', () => {
-    toggleIconPacks();
-});
-
-// Folder themes
-const changeFolderThemeCommand = vscode.commands.registerCommand('material-icon-theme.changeFolderTheme', () => {
-    changeFolderTheme();
-});
-
-// Folder color
-const toggleFolderColorCommand = vscode.commands.registerCommand('material-icon-theme.changeFolderColor', () => {
-    changeFolderColor();
-});
-
-// Reset config
-const restoreDefaultConfigCommand = vscode.commands.registerCommand('material-icon-theme.restoreDefaultConfig', () => {
-    restoreDefaultConfig();
-});
-
-// Toggle the arrows near the folder icons
-const hidesExplorerArrowsCommand = vscode.commands.registerCommand('material-icon-theme.hidesExplorerArrows', () => {
-    toggleFolderArrows();
-});
-
-// Change the opacity of the icons
-const changeOpacityCommand = vscode.commands.registerCommand('material-icon-theme.opacity', () => {
-    changeOpacity();
-});
-
-export const commands = [
-    activateThemeCommand,
-    toggleIconPacksCommand,
-    changeFolderThemeCommand,
-    toggleFolderColorCommand,
-    restoreDefaultConfigCommand,
-    hidesExplorerArrowsCommand,
-    changeOpacityCommand
-];