+++ src/commands/opacity.ts
@@ -4,20 +4,24 @@ import * as helpers from './../helpers';
 import * as i18n from './../i18n';
 
 /** Command to toggle the folder icons. */
-export const changeOpacity = () => {
-    return getCurrentOpacityValue()
-        .then(showInput)
-        .catch(err => console.log(err));
+export const changeOpacity = async () => {
+    try {
+        const currentOpacityValue = getCurrentOpacityValue();
+        const response = Number(await showInput(currentOpacityValue));
+        return setOpacityConfig(response);
+    } catch (error) {
+        console.error(error);
+    }
 };
 
 /** Show input to enter the opacity value. */
 const showInput = (opacity: number) => {
-    vscode.window.showInputBox({
+    return vscode.window.showInputBox({
         placeHolder: i18n.translate('opacity.inputPlaceholder'),
         ignoreFocusOut: true,
         value: String(opacity),
         validateInput: validateOpacityInput
-    }).then(value => setOpacityConfig(+value));
+    });
 };
 
 /** Validate the opacity value which was inserted by the user. */
@@ -29,15 +33,15 @@ const validateOpacityInput = (opacityInput: string) => {
 };
 
 /** Get the current value of the opacity of the icons. */
-export const getCurrentOpacityValue = (): Promise<number> => {
+export const getCurrentOpacityValue = (): number => {
     const defaultOptions = getDefaultIconOptions();
-    return helpers.getMaterialIconsJSON().then((config) =>
-        config.options.opacity === undefined ?
-            defaultOptions.opacity : config.options.opacity);
+    const config = helpers.getMaterialIconsJSON();
+    return config.options.opacity === undefined ?
+        defaultOptions.opacity : config.options.opacity;
 };
 
 const setOpacityConfig = (opacity: number) => {
     if (opacity !== undefined) {
-        helpers.setThemeConfig('opacity', opacity, true);
+        return helpers.setThemeConfig('opacity', opacity, true);
     }
 };