+++ src/icons/generator/constants.ts
@@ -1,7 +1,7 @@
 /**
  * Path where the icons are located.
  */
-export const iconFolderPath: string = './../../icons/';
+export const iconFolderPath: string = './../icons/';
 
 /**
  * File name of the JSON file that will be generated to the out folder.